requirements.txt
asgiref==3.2.2
asn1crypto==1.0.1
attrs==19.2.0
autobahn==19.10.1
Automat==0.7.0
certifi==2019.9.11
cffi==1.12.3
channels==2.3.0
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
dj-database-url==0.5.0
Django==2.2.5
django-heroku==0.3.1
django-rest-swagger==2.1.2
djangorestframework==3.10.3
djangorestframework-jwt==1.11.0
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
psycopg2==2.8.3
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
pytz==2019.2
requests==2.22.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
Twisted==19.7.0
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.6
whitenoise==4.1.4
zope.interface==4.6.0
django-cors-headers
channels_redis
django-rest-passwordreset
django-filter
Pillow